Update Hindi translations in strings.xml
### Summary

- Updated 7 existing hindi strings.
---
Note: - Ready for more bulk changes in hindi string in the future.

### Special note:- 

I am very happy that you came back with new motivation for continuing the project from Rimusic to Riplay. At that time when Rimsuic not archive and no drama of clone Rimusic and Rimusic discord channel issue exists, I was only able to report about issues or help someone in questions for generic cases but after 8-9 months which gived a great experience and knowledge, I am happy to say that it is going to be my first contribution in **RiPlay** as a contributor. For now I am only able to do little changes as I am in my sleep time 😅 but I am glad I did that.

   So, well I am looking forward for great journey now as being a part of it. I hope you will continue to do the work well and be an inspiration for some small learners like us. 
Thansk